requirements.txt
django
django-crispy-forms
phonenumbers
django-phonenumber-field
django-gst-field